how LSTMs work and code?
LSTM is nothing but the long short term memory, it is an artificial recurrent neural network used in the field of deep learning, also LSTM's are a special kind of RNN, capable of learning long term dependencies. These networks are based on time series data which are well suited for classifying, processing, and making predictions. Also, the development of these was done for dealing with the vanishing gradient problem that can be encountered when training traditional RNNs.